Single file per integration test. Remove duplicate cell_list test and include a doubly nested cell to the other test.
